added new media, added more gui options
Former-commit-id: 9ac2ed4f80a6be4613db69e5678355618d2a4094

+++ src/0.1.04/framework/vka/specialized_utility/compute.cpp
@@ -210,7 +210,7 @@ ComputeCmd getCmdPerlinNoise(Image target, PerlinNoiseArgs args)
 		float    minVal;
 		float    maxVal;
 		float    frequency;
-		uint32_t falloffAtEdge;
+		float	falloffAtEdge;
 		uint32_t blendOp;
 	} pc;
 	pc.seed         = args.seed;
